Move scheduled tasks to a seperate file
