+++ app/__init__.py
@@ -6,6 +6,10 @@ import warnings
 from flask.exthook import ExtDeprecationWarning
 from forex_python.converter import CurrencyCodes
 from pytz import utc
+
+from app.helpers.scheduled_jobs import send_mail_to_expired_orders, empty_trash, send_after_event_mail, \
+    send_event_fee_notification, send_event_fee_notification_followup
+
 warnings.simplefilter('ignore', ExtDeprecationWarning)
 # Keep it before flask extensions are imported
 import arrow
@@ -51,7 +55,6 @@ from apscheduler.schedulers.background import BackgroundScheduler
 from app.helpers.data import DataManager, delete_from_db
 from app.helpers.helpers import send_after_event
 from app.helpers.cache import cache
-from sqlalchemy_continuum import transaction_class
 from helpers.helpers import send_email_for_expired_orders
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -362,80 +365,17 @@ def integrate_socketio():
     return dict(integrate_socketio=integrate)
 
 
-def send_after_event_mail():
-    with app.app_context():
-        events = Event.query.all()
-        for event in events:
-            upcoming_events = DataGetter.get_upcoming_events(event.id)
-            organizers = DataGetter.get_user_event_roles_by_role_name(
-                event.id, 'organizer')
-            speakers = DataGetter.get_user_event_roles_by_role_name(event.id,
-                                                                    'speaker')
-            if datetime.now() > event.end_time:
-                for speaker in speakers:
-                    send_after_event(speaker.user.email, event.id,
-                                     upcoming_events)
-                for organizer in organizers:
-                    send_after_event(organizer.user.email, event.id,
-                                     upcoming_events)
-
-
-# logging.basicConfig()
-sched = BackgroundScheduler(timezone=utc)
-sched.add_job(send_after_event_mail,
-              'cron',
-              day_of_week='mon-fri',
-              hour=5,
-              minute=30)
-
-
-# sched.start()
-
-def send_mail_to_expired_orders():
-    with app.app_context():
-        orders = DataGetter.get_expired_orders()
-        for order in orders:
-            send_email_for_expired_orders('adityavyas17@gmail.com', order.event.name, order.get_invoice_number(),
-                                          url_for('ticketing.view_order_after_payment',
-                                                  order_identifier=order.identifier, _external=True))
-
-order_sched = BackgroundScheduler(timezone=utc)
-order_sched.add_job(send_mail_to_expired_orders,
-                    'interval',
-                    hours=5)
-
-
-order_sched.start()
-
-def empty_trash():
-    with app.app_context():
-        events = Event.query.filter_by(in_trash=True)
-        users = User.query.filter_by(in_trash=True)
-        sessions = Session.query.filter_by(in_trash=True)
-        for event in events:
-            if datetime.now() - event.trash_date >= timedelta(days=30):
-                DataManager.delete_event(event.id)
-
-        for user in users:
-            if datetime.now() - user.trash_date >= timedelta(days=30):
-                transaction = transaction_class(Event)
-                transaction.query.filter_by(user_id=user.id).delete()
-                delete_from_db(user, "User deleted permanently")
-
-        for session_ in sessions:
-            if datetime.now() - session_.trash_date >= timedelta(days=30):
-                delete_from_db(session_, "Session deleted permanently")
-
-
-trash_sched = BackgroundScheduler(timezone=utc)
-trash_sched.add_job(
-    empty_trash, 'cron',
-    day_of_week='mon-fri',
-    hour=5, minute=30)
-trash_sched.start()
+scheduler = BackgroundScheduler(timezone=utc)
+scheduler.add_job(send_mail_to_expired_orders, 'interval', hours=5)
+scheduler.add_job(empty_trash, 'cron', day_of_week='mon-fri', hour=5, minute=30)
+scheduler.add_job(send_after_event_mail, 'cron', day_of_week='mon-fri', hour=5, minute=30)
+scheduler.add_job(send_event_fee_notification, 'cron', day=1)
+scheduler.add_job(send_event_fee_notification_followup, 'cron', day=15)
+scheduler.start()
 
 # Flask-SocketIO integration
 
+socketio = None
 if current_app.config.get('INTEGRATE_SOCKETIO', False):
     from eventlet import monkey_patch
     from flask_socketio import SocketIO, emit, join_room
